+++ tools/byte_converter/src/hooks/use-value.ts
@@ -1,9 +1,15 @@
 import { useCallback, useMemo } from "preact/hooks";
 import { Encoding, Unit } from "../types";
+import { clamp, padL } from "../utils";
 
 export const Boundaries = {
   [Unit.Byte]: { min: 0, max: 255 },
   [Unit.Word]: { min: 0, max: 65535 },
+} as const;
+
+export const SignedBoundaries = {
+  [Unit.Byte]: { min: -129, max: 127 },
+  [Unit.Word]: { min: -32768, max: 32767 },
 } as const;
 
 export const Chars = {
@@ -34,7 +40,8 @@ export const Radix = {
 export function useValue(
   integer: number,
   encoding: Encoding,
-  unit: Unit
+  unit: Unit,
+  signedEnabled: boolean
 ): [
   string,
   {
@@ -45,28 +52,50 @@ export function useValue(
     validChar: (char: string) => boolean;
   }
 ] {
+  const toSigned = useCallback(
+    (n: number): number => {
+      if (!signedEnabled) return n;
+      if (n <= SignedBoundaries[unit].max) return n;
+      return -(2 * (SignedBoundaries[unit].max + 1) - n);
+    },
+    [signedEnabled, unit]
+  );
+
+  const fromSigned = useCallback(
+    (n: number): number => {
+      if (!signedEnabled) return n;
+      if (n >= 0) return n;
+      return 2 * (SignedBoundaries[unit].max + 1) + n;
+    },
+    [signedEnabled, unit]
+  );
+
   const value = useMemo(() => {
     const { min, max } = Boundaries[unit];
     const length = Length[unit][encoding];
-    const n = Math.max(Math.min(integer, max), min)
-      .toString(Radix[encoding])
-      .toUpperCase();
-    return `${"0".repeat(length - n.length)}${n}`;
-  }, [encoding, integer, unit]);
+    const signed = toSigned(clamp(integer, min, max));
+    const signedStr = signed.toString(Radix[encoding]).toUpperCase();
+    return signed < 0
+      ? `-${padL(signedStr.substring(1), length, "0")}`
+      : signedEnabled
+      ? ` ${padL(signedStr, length, "0")}`
+      : padL(signedStr, length, "0");
+  }, [encoding, integer, toSigned, unit]);
 
   const parse = useCallback(
     (
       maybeValue: string,
       bounds?: { min?: number; max?: number }
     ): number | undefined => {
-      const i = Number.parseInt(maybeValue, Radix[encoding]);
-      const { min, max } = Boundaries[unit];
+      const signed = Number.parseInt(maybeValue, Radix[encoding]);
+      if (Number.isNaN(signed)) return undefined;
+      const { min, max } = signedEnabled
+        ? SignedBoundaries[unit]
+        : Boundaries[unit];
       bounds = { ...Boundaries[unit], ...bounds };
-      if (i < min) return bounds.min;
-      if (i > max) return bounds.max;
-      return Number.isNaN(i) ? undefined : i;
+      return fromSigned(clamp(signed, min, max));
     },
-    [encoding, unit]
+    [encoding, fromSigned, signedEnabled, unit]
   );
 
   const validChar = useCallback(